store/wizard: extract content selectors to subslice
Moves all content-related selectors from slice.ts into
content/selectors.ts, including the derived selectAllRepositoryIds.
