+++ src/store/slices/wizard/content/index.ts
@@ -1,2 +1,3 @@
+export * from './selectors';
 export { initialState as contentState } from './state';
 export * from './types';
