+++ src/store/slices/wizard/slice.ts
@@ -72,50 +72,6 @@ export const initialState: WizardState = {
   },
 };
 
-export const selectUseLatest = (state: RootState) => {
-  return state.wizard.content.snapshotting.useLatest;
-};
-
-export const selectSnapshotDate = (state: RootState) => {
-  return state.wizard.content.snapshotting.snapshotDate;
-};
-
-export const selectTemplate = (state: RootState) => {
-  return state.wizard.content.snapshotting.template;
-};
-
-export const selectTemplateName = (state: RootState) => {
-  return state.wizard.content.snapshotting.templateName;
-};
-
-export const selectCustomRepositories = (state: RootState) => {
-  return state.wizard.content.repositories.customRepositories;
-};
-
-export const selectPayloadRepositories = (state: RootState) => {
-  return state.wizard.content.repositories.payloadRepositories;
-};
-
-export const selectRecommendedRepositories = (state: RootState) => {
-  return state.wizard.content.repositories.recommendedRepositories;
-};
-
-export const selectRedHatRepositories = (state: RootState) => {
-  return state.wizard.content.repositories.redHatRepositories;
-};
-
-export const selectPackages = (state: RootState) => {
-  return state.wizard.content.packages;
-};
-
-export const selectModules = (state: RootState) => {
-  return state.wizard.content.enabledModules;
-};
-
-export const selectPackageGroups = (state: RootState) => {
-  return state.wizard.content.groups;
-};
-
 export const selectServices = (state: RootState) => {
   return state.wizard.system.services;
 };
@@ -175,10 +131,6 @@ export const selectFirewall = (state: RootState) => {
   return state.wizard.system.firewall;
 };
 
-export const selectVerifiedLocaleLangpacks = (state: RootState) => {
-  return state.wizard.content.verifiedLocaleLangpacks;
-};
-
 const extractLanguageCode = (locale: string): string | undefined => {
   const [regionPart] = locale.split('.');
   const [languageCode] = regionPart.split('_');
@@ -219,23 +171,6 @@ export const selectFirewallEnabled = createSelector(
     if (firewall.services.disabled.length > 0) return true;
     return false;
   },
-);
-
-// Derived selector for getting all repositories
-export const selectAllRepositoryIds = createSelector(
-  [
-    selectCustomRepositories,
-    selectPayloadRepositories,
-    selectRecommendedRepositories,
-  ],
-  (custom, payload, recommended) =>
-    Array.from(
-      new Set([
-        ...custom.map(({ id }) => id),
-        ...payload.flatMap(({ id }) => (id ? [id] : [])),
-        ...recommended.flatMap(({ uuid }) => (uuid ? [uuid] : [])),
-      ]),
-    ),
 );
 
 export const wizardSlice = createSlice({
